@@ -1,7 +1,8 @@
+"use client";
 import { useEffect, useState } from "react";
 import { Results } from "./Results";
 import { Pagination } from "./Pagination";
-import { useRouter } from "next/router";
+import { useRouter, usePathname } from "next/navigation";
 import queryString from "query-string";
 
 export const BlogSearch = () => {
@@ -9,6 +10,7 @@ export const BlogSearch = () => {
   const [totalResults, setTotalResults] = useState(0);
   const pageSize = 3;
   const router = useRouter();
+  const pathName = usePathname();
 
   const search = async () => {
     const { page } = queryString.parse(window.location.search);
@@ -24,14 +26,7 @@ export const BlogSearch = () => {
   };
 
   const handlePageClick = async (pageNum) => {
-    await router.push(
-      `${router.query.slug.join("/")}?page=${[pageNum]}`,
-      null,
-      {
-        shallow: true,
-      }
-    );
-    search();
+    router.push(`${pathName.query.slug.join("/")}?page=${[pageNum]}`);
   };
 
   useEffect(() => {
