@@ -2,7 +2,7 @@ import Link from "next/link";
 
 export const ButtonLink = ({ destination, label }) => {
   return (
-    <Link href={destination} className="btn">
+    <Link href={`${destination}`} className="btn">
       {label}
     </Link>
   );
