@@ -3,13 +3,14 @@ module.exports = {
   content: [
     "./pages/**/*.{js,ts,jsx,tsx}",
     "./components/**/*.{js,ts,jsx,tsx}",
+    "./app/**/*.{js,ts,jsx,tsx}",
     "./utils/**/*.js",
   ],
   theme: {
     extend: {
       fontFamily: {
-        heading: ["Monaco"],
-        body: ["Poppins", "sans-serif"],
+        heading: "var(--font-aboreto)",
+        body: "var(--font-poppins)",
       },
       colors: {
         blue: {
